@@ -3,64 +3,56 @@
   "version": "0.3.0",
   "author": "mutoe <mutoe@foxmail.com>",
   "license": "MIT",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/mutoe/cms-page-builder.git"
   },
   "scripts": {
     "prepare": "husky install",
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public",
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "lint": "eslint 'src/**/*.{ts,js}' && svelte-check",
     "test": "jest src",
     "cy": "cypress",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "version": "yarn changelog && git add CHANGELOG.md"
   },
-  "dependencies": {
-    "sirv-cli": "^1.0.8"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-image": "^2.0.6",
-    "@rollup/plugin-node-resolve": "^11.2.0",
-    "@rollup/plugin-typescript": "^8.2.0",
-    "@tsconfig/svelte": "^1.0.10",
-    "postcss": "^8.2.6",
-    "rollup": "^2.40.0",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.1.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "sass": "^1.32.8",
-    "svelte": "^3.35.0",
-    "svelte-check": "^1.2.1",
-    "svelte-jester": "^1.3.0",
-    "svelte-preprocess": "^4.6.9",
-    "tslib": "^2.1.0",
-    "typescript": "^4.2.2"
+    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.17",
+    "@tsconfig/svelte": "^2.0.1",
+    "sass": "^1.37.5",
+    "svelte": "^3.42.1",
+    "svelte-check": "^2.2.4",
+    "svelte-preprocess": "^4.7.4",
+    "tslib": "^2.3.1",
+    "typescript": "^4.3.5",
+    "vite": "^2.5.0",
+    "vite-tsconfig-paths": "^3.3.13"
   },
   "optionalDependencies": {
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
-    "@testing-library/cypress": "^7.0.4",
-    "@testing-library/jest-dom": "^5.11.9",
+    "@testing-library/cypress": "^8.0.0",
+    "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/svelte": "^3.0.3",
-    "@types/jest": "^26.0.20",
-    "@typescript-eslint/eslint-plugin": "^4.16.1",
-    "@typescript-eslint/parser": "^4.16.1",
-    "cypress": "^6.6.0",
-    "eslint": "^7.21.0",
+    "@types/jest": "^26.0.24",
+    "@typescript-eslint/eslint-plugin": "^4.29.1",
+    "@typescript-eslint/parser": "^4.29.1",
+    "cypress": "^8.2.0",
+    "eslint": "^7.32.0",
     "eslint-config-standard-with-typescript": "^20.0.0",
-    "eslint-plugin-cypress": "^2.11.2",
-    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-cypress": "^2.11.3",
+    "eslint-plugin-import": "^2.24.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.3.1",
-    "husky": "^6.0.0",
+    "husky": "^7.0.1",
     "jest": "^26.6.3",
-    "lint-staged": "^10.5.4",
-    "ts-jest": "^26.5.2"
+    "lint-staged": "^11.1.2",
+    "svelte-jester": "^1.3.0",
+    "ts-jest": "^26.5.6"
   },
   "lint-staged": {
     "src/**/*.{js,ts}": "eslint --fix"
